Demo Displays
Add demo data displays for the presentation
