+++ openremote/src/lib/extractData.js
@@ -1,0 +1,7 @@
+const _ = require('lodash');
+
+module.exports = {
+    extractData: (array) => {
+        return _.map(array, 'value');
+    }
+}
