+++ openremote/src/styles/powergraph.component.css
@@ -2,5 +2,5 @@
     position: relative;
     margin-top: 1rem;
     height: 10vh;
-    width: 30vw;
+    width: 50vw;
 }
